Lesson 3 - task - 3 - Добавить Android View UI элементы - Подключить layout (xml файл) к фрагменту - Подключил app bar и добавил кнопку назад - Исправил навигацию, из предыдущего урока
