+++ app/src/main/java/com/example/crimedigital/premierleaguefixtures/screens/Container.kt
@@ -1,0 +1,13 @@
+package com.example.crimedigital.premierleaguefixtures.screens
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.crimedigital.premierleaguefixtures.R
+
+class Container : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.container)
+    }
+}
