+++ app/src/main/java/com/example/crimedigital/premierleaguefixtures/screens/MatchListFragment.kt
@@ -1,34 +1,33 @@
-package com.example.crimedigital.premierleaguefixtures.lesson_2
+package com.example.crimedigital.premierleaguefixtures.screens
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
 import com.example.crimedigital.premierleaguefixtures.R
 import com.example.crimedigital.premierleaguefixtures.databinding.FragmentMatchListBinding
 
-
 class MatchListFragment : Fragment() {
-    private lateinit var binding: FragmentMatchListBinding
+    private var _binding: FragmentMatchListBinding? = null
+    private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
+        _binding = FragmentMatchListBinding.inflate(inflater, container, false)
 
-        binding = FragmentMatchListBinding.inflate(inflater, container, false)
         binding.buttonMatchDetails.setOnClickListener {
-            val fragment = MatchDetailsFragment()
-            val transaction = requireActivity().supportFragmentManager.beginTransaction()
-            transaction.replace(R.id.fragmentContainerView, fragment)
-                .commit()
+            findNavController().navigate(R.id.action_matchListFragment_to_matchDetailsFragment)
         }
-
-
 
         return binding.root
     }
 
-
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
 }
