+++ app/src/main/res/layout/fragment_match_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/red"
-    tools:context=".lesson_2.MatchListFragment">
+    tools:context=".screens.MatchListFragment">
 
     <TextView
         android:id="@+id/tv_match_list"
