--- app/src/main/res/values-v23/themes.xml
@@ -1,8 +1,0 @@
-<resources >
-  <style name="Theme.PremierLeagueFixtures" parent="Base.Theme.PremierLeagueFixtures">
-
-    <item name="android:navigationBarColor">@android:color/transparent</item>
-    <item name="android:statusBarColor">@android:color/transparent</item>
-    <item name="android:windowLightStatusBar">?attr/isLightTheme</item>
-  </style>
-</resources>
